buble

+++ 0-bubble_sort.c
@@ -9,7 +9,7 @@
 
 void bubble_sort(int *array, size_t size)
 {
-	size_t i, j, swap;
+	unsigned int i, j, swap;
 
 	for (i = 0; i < size - 1; i++)
 	{
